+++ src/components/nodes/cards/stanford-course/stanford-course-card.tsx
@@ -14,7 +14,7 @@ const StanfordCourseCard = ({node, headingLevel, ...props}: Props) => {
   return (
     <ImageCard {...props} aria-labelledby={node.id} isArticle>
       <Heading className="type-3 order-last" id={node.id}>
-        <Link href={node.path}>{node.title}</Link>
+        <Link href={node.path || "#"}>{node.title}</Link>
       </Heading>
       <div className="order-first flex gap-5">
         {node.suCourseSubject && (