+++ src/components/nodes/cards/stanford-news/stanford-news-card.tsx
@@ -28,7 +28,7 @@ const StanfordNewsCard = ({node, headingLevel, ...props}: Props) => {
     <ImageCard {...props} aria-labelledby={node.id} imageUrl={image?.url} imageAlt={image?.alt} isArticle>
       <div className="flex flex-col">
         <Heading className="type-3 [&_a]:text-black" id={node.id}>
-          <Link href={node.suNewsSource?.url || node.path}>{node.title}</Link>
+          <Link href={node.suNewsSource?.url || node.path || "#"}>{node.title}</Link>
         </Heading>
 
         {publishDate && <div className="order-first">{publishDate}</div>}